telegram_ui/requirements.txt
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0
langgraph-sdk==0.0.26
httpx==0.24.1 